+++ src/fourc/fmf/Broker.cxx
@@ -65,361 +65,361 @@ Broker::Broker() :
   mgmtPublish(false),
   uptime(std::chrono::nanoseconds::min()) {}
 
-uint32_t Broker::getAbandoned() const {
+uint64_t Broker::getAbandoned() const {
   return abandoned;
 }
 
-Broker& Broker::setAbandoned(const uint32_t value) {
+Broker& Broker::setAbandoned(const uint64_t value) {
   abandoned = value;
 
   return *this;
 }
 
-uint32_t Broker::getAbandonedViaAlt() const {
+uint64_t Broker::getAbandonedViaAlt() const {
   return abandonedViaAlt;
 }
 
-Broker& Broker::setAbandonedViaAlt(const uint32_t value) {
+Broker& Broker::setAbandonedViaAlt(const uint64_t value) {
   abandonedViaAlt = value;
 
   return *this;
 }
 
-uint32_t Broker::getAcquires() const {
+uint64_t Broker::getAcquires() const {
   return acquires;
 }
 
-Broker& Broker::setAcquires(const uint32_t value) {
+Broker& Broker::setAcquires(const uint64_t value) {
   acquires = value;
 
   return *this;
 }
 
-uint32_t Broker::getByteDepth() const {
+uint64_t Broker::getByteDepth() const {
   return byteDepth;
 }
 
-Broker& Broker::setByteDepth(const uint32_t value) {
+Broker& Broker::setByteDepth(const uint64_t value) {
   byteDepth = value;
 
   return *this;
 }
 
-uint32_t Broker::getByteFtdDepth() const {
+uint64_t Broker::getByteFtdDepth() const {
   return byteFtdDepth;
 }
 
-Broker& Broker::setByteFtdDepth(const uint32_t value) {
+Broker& Broker::setByteFtdDepth(const uint64_t value) {
   byteFtdDepth = value;
 
   return *this;
 }
 
-uint32_t Broker::getByteFtdDequeues() const {
+uint64_t Broker::getByteFtdDequeues() const {
   return byteFtdDequeues;
 }
 
-Broker& Broker::setByteFtdDequeues(const uint32_t value) {
+Broker& Broker::setByteFtdDequeues(const uint64_t value) {
   byteFtdDequeues = value;
 
   return *this;
 }
 
-uint32_t Broker::getByteFtdEnqueues() const {
+uint64_t Broker::getByteFtdEnqueues() const {
   return byteFtdEnqueues;
 }
 
-Broker& Broker::setByteFtdEnqueues(const uint32_t value) {
+Broker& Broker::setByteFtdEnqueues(const uint64_t value) {
   byteFtdEnqueues = value;
 
   return *this;
 }
 
-uint32_t Broker::getBytePersistDequeues() const {
+uint64_t Broker::getBytePersistDequeues() const {
   return bytePersistDequeues;
 }
 
-Broker& Broker::setBytePersistDequeues(const uint32_t value) {
+Broker& Broker::setBytePersistDequeues(const uint64_t value) {
   bytePersistDequeues = value;
 
   return *this;
 }
 
-uint32_t Broker::getBytePersistEnqueues() const {
+uint64_t Broker::getBytePersistEnqueues() const {
   return bytePersistEnqueues;
 }
 
-Broker& Broker::setBytePersistEnqueues(const uint32_t value) {
+Broker& Broker::setBytePersistEnqueues(const uint64_t value) {
   bytePersistEnqueues = value;
 
   return *this;
 }
 
-uint32_t Broker::getByteTotalDequeues() const {
+uint64_t Broker::getByteTotalDequeues() const {
   return byteTotalDequeues;
 }
 
-Broker& Broker::setByteTotalDequeues(const uint32_t value) {
+Broker& Broker::setByteTotalDequeues(const uint64_t value) {
   byteTotalDequeues = value;
 
   return *this;
 }
 
-uint32_t Broker::getByteTotalEnqueues() const {
+uint64_t Broker::getByteTotalEnqueues() const {
   return byteTotalEnqueues;
 }
 
-Broker& Broker::setByteTotalEnqueues(const uint32_t value) {
+Broker& Broker::setByteTotalEnqueues(const uint64_t value) {
   byteTotalEnqueues = value;
 
   return *this;
 }
 
-uint32_t Broker::getByteTxnDequeues() const {
+uint64_t Broker::getByteTxnDequeues() const {
   return byteTxnDequeues;
 }
 
-Broker& Broker::setByteTxnDequeues(const uint32_t value) {
+Broker& Broker::setByteTxnDequeues(const uint64_t value) {
   byteTxnDequeues = value;
 
   return *this;
 }
 
-uint32_t Broker::getByteTxnEnqueues() const {
+uint64_t Broker::getByteTxnEnqueues() const {
   return byteTxnEnqueues;
 }
 
-Broker& Broker::setByteTxnEnqueues(const uint32_t value) {
+Broker& Broker::setByteTxnEnqueues(const uint64_t value) {
   byteTxnEnqueues = value;
 
   return *this;
 }
 
-uint32_t Broker::getConnBacklog() const {
+uint16_t Broker::getConnBacklog() const {
   return connBacklog;
 }
 
-Broker& Broker::setConnBacklog(const uint32_t value) {
+Broker& Broker::setConnBacklog(const uint16_t value) {
   connBacklog = value;
 
   return *this;
 }
 
-uint32_t Broker::getDiscardsLvq() const {
+uint64_t Broker::getDiscardsLvq() const {
   return discardsLvq;
 }
 
-Broker& Broker::setDiscardsLvq(const uint32_t value) {
+Broker& Broker::setDiscardsLvq(const uint64_t value) {
   discardsLvq = value;
 
   return *this;
 }
 
-uint32_t Broker::getDiscardsNoRoute() const {
+uint64_t Broker::getDiscardsNoRoute() const {
   return discardsNoRoute;
 }
 
-Broker& Broker::setDiscardsNoRoute(const uint32_t value) {
+Broker& Broker::setDiscardsNoRoute(const uint64_t value) {
   discardsNoRoute = value;
 
   return *this;
 }
 
-uint32_t Broker::getDiscardsOverflow() const {
+uint64_t Broker::getDiscardsOverflow() const {
   return discardsOverflow;
 }
 
-Broker& Broker::setDiscardsOverflow(const uint32_t value) {
+Broker& Broker::setDiscardsOverflow(const uint64_t value) {
   discardsOverflow = value;
 
   return *this;
 }
 
-uint32_t Broker::getDiscardsPurge() const {
+uint64_t Broker::getDiscardsPurge() const {
   return discardsPurge;
 }
 
-Broker& Broker::setDiscardsPurge(const uint32_t value) {
+Broker& Broker::setDiscardsPurge(const uint64_t value) {
   discardsPurge = value;
 
   return *this;
 }
 
-uint32_t Broker::getDiscardsRing() const {
+uint64_t Broker::getDiscardsRing() const {
   return discardsRing;
 }
 
-Broker& Broker::setDiscardsRing(const uint32_t value) {
+Broker& Broker::setDiscardsRing(const uint64_t value) {
   discardsRing = value;
 
   return *this;
 }
 
-uint32_t Broker::getDiscardsSubscriber() const {
+uint64_t Broker::getDiscardsSubscriber() const {
   return discardsSubscriber;
 }
 
-Broker& Broker::setDiscardsSubscriber(const uint32_t value) {
+Broker& Broker::setDiscardsSubscriber(const uint64_t value) {
   discardsSubscriber = value;
 
   return *this;
 }
 
-uint32_t Broker::getDiscardsTtl() const {
+uint64_t Broker::getDiscardsTtl() const {
   return discardsTtl;
 }
 
-Broker& Broker::setDiscardsTtl(const uint32_t value) {
+Broker& Broker::setDiscardsTtl(const uint64_t value) {
   discardsTtl = value;
 
   return *this;
 }
 
-uint32_t Broker::getMaxConns() const {
+uint16_t Broker::getMaxConns() const {
   return maxConns;
 }
 
-Broker& Broker::setMaxConns(const uint32_t value) {
+Broker& Broker::setMaxConns(const uint16_t value) {
   maxConns = value;
 
   return *this;
 }
 
-uint32_t Broker::getMgmtPubInterval() const {
+uint16_t Broker::getMgmtPubInterval() const {
   return mgmtPubInterval;
 }
 
-Broker& Broker::setMgmtPubInterval(const uint32_t value) {
+Broker& Broker::setMgmtPubInterval(const uint16_t value) {
   mgmtPubInterval = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgDepth() const {
+uint64_t Broker::getMsgDepth() const {
   return msgDepth;
 }
 
-Broker& Broker::setMsgDepth(const uint32_t value) {
+Broker& Broker::setMsgDepth(const uint64_t value) {
   msgDepth = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgFtdDepth() const {
+uint64_t Broker::getMsgFtdDepth() const {
   return msgFtdDepth;
 }
 
-Broker& Broker::setMsgFtdDepth(const uint32_t value) {
+Broker& Broker::setMsgFtdDepth(const uint64_t value) {
   msgFtdDepth = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgFtdDequeues() const {
+uint64_t Broker::getMsgFtdDequeues() const {
   return msgFtdDequeues;
 }
 
-Broker& Broker::setMsgFtdDequeues(const uint32_t value) {
+Broker& Broker::setMsgFtdDequeues(const uint64_t value) {
   msgFtdDequeues = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgFtdEnqueues() const {
+uint64_t Broker::getMsgFtdEnqueues() const {
   return msgFtdEnqueues;
 }
 
-Broker& Broker::setMsgFtdEnqueues(const uint32_t value) {
+Broker& Broker::setMsgFtdEnqueues(const uint64_t value) {
   msgFtdEnqueues = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgPersistDequeues() const {
+uint64_t Broker::getMsgPersistDequeues() const {
   return msgPersistDequeues;
 }
 
-Broker& Broker::setMsgPersistDequeues(const uint32_t value) {
+Broker& Broker::setMsgPersistDequeues(const uint64_t value) {
   msgPersistDequeues = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgPersistEnqueues() const {
+uint64_t Broker::getMsgPersistEnqueues() const {
   return msgPersistEnqueues;
 }
 
-Broker& Broker::setMsgPersistEnqueues(const uint32_t value) {
+Broker& Broker::setMsgPersistEnqueues(const uint64_t value) {
   msgPersistEnqueues = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgTotalDequeues() const {
+uint64_t Broker::getMsgTotalDequeues() const {
   return msgTotalDequeues;
 }
 
-Broker& Broker::setMsgTotalDequeues(const uint32_t value) {
+Broker& Broker::setMsgTotalDequeues(const uint64_t value) {
   msgTotalDequeues = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgTotalEnqueues() const {
+uint64_t Broker::getMsgTotalEnqueues() const {
   return msgTotalEnqueues;
 }
 
-Broker& Broker::setMsgTotalEnqueues(const uint32_t value) {
+Broker& Broker::setMsgTotalEnqueues(const uint64_t value) {
   msgTotalEnqueues = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgTxnDequeues() const {
+uint64_t Broker::getMsgTxnDequeues() const {
   return msgTxnDequeues;
 }
 
-Broker& Broker::setMsgTxnDequeues(const uint32_t value) {
+Broker& Broker::setMsgTxnDequeues(const uint64_t value) {
   msgTxnDequeues = value;
 
   return *this;
 }
 
-uint32_t Broker::getMsgTxnEnqueues() const {
+uint64_t Broker::getMsgTxnEnqueues() const {
   return msgTxnEnqueues;
 }
 
-Broker& Broker::setMsgTxnEnqueues(const uint32_t value) {
+Broker& Broker::setMsgTxnEnqueues(const uint64_t value) {
   msgTxnEnqueues = value;
 
   return *this;
 }
 
-uint32_t Broker::getQueueCount() const {
+uint64_t Broker::getQueueCount() const {
   return queueCount;
 }
 
-Broker& Broker::setQueueCount(const uint32_t value) {
+Broker& Broker::setQueueCount(const uint64_t value) {
   queueCount = value;
 
   return *this;
 }
 
-uint32_t Broker::getReleases() const {
+uint64_t Broker::getReleases() const {
   return releases;
 }
 
-Broker& Broker::setReleases(const uint32_t value) {
+Broker& Broker::setReleases(const uint64_t value) {
   releases = value;
 
   return *this;
 }
 
-uint32_t Broker::getReroutes() const {
+uint64_t Broker::getReroutes() const {
   return reroutes;
 }
 
-Broker& Broker::setReroutes(const uint32_t value) {
+Broker& Broker::setReroutes(const uint64_t value) {
   reroutes = value;
 
   return *this;
